feat: finish crud and upload csv

+++ src/routes.js
@@ -32,6 +32,18 @@ export const routes = [
     handler: (req, res) => {
       const { title, description } = req.body
 
+      if (!title) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: 'title is required' }))
+      }
+
+      if (!description) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: 'description is required' }))
+      }
+
       const task = {
         id: randomUUID(),
         title,
@@ -53,9 +65,22 @@ export const routes = [
       const { id } = req.params
       const { title, description } = req.body
 
+      if (!title || !description) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: 'title or description are required' }))
+      }
+
+      const [task] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
       const data = {
-        title,
-        description,
+        title: title ?? task.title,
+        description: description ?? task.description,
+        updated_at: new Date(),
       }
 
       database.update('tasks', id, data)
@@ -69,7 +94,16 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params
 
-      database.complete('tasks', id)
+      const [task] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
+
+      const isTaskCompleted = !!task.completed_at
+      const completed_at = isTaskCompleted ? null : new Date()
+
+      database.update('tasks', id, { completed_at })
 
       return res.writeHead(204).end()
     },
@@ -79,6 +113,12 @@ export const routes = [
     path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
       const { id } = req.params
+
+      const [task] = database.select('tasks', { id })
+
+      if (!task) {
+        return res.writeHead(404).end()
+      }
 
       database.delete('tasks', id)
